colour-service/requirements.txt
fastapi
uvicorn[standard]
pillow
openai
python-multipart